Struct wu_manber::TwoByteWM

pub struct TwoByteWM {
    // some fields omitted
}

TwoByteWM stores the precomputed tables needed for a two-byte-wide implementation of the Wu-Manber algorithm.

"Two-byte-wide" means that the search phase in the Wu-Manber algorithm uses spans of two bytes to look for potential matches. This is suitable for moderately sized sets of needles; if there are too many needles then it might be faster to use spans of three bytes (but that isn't yet implemented by this crate).

Methods

impl TwoByteWM

fn new<I, P>(needles: I) -> TwoByteWM where P: AsRef<[u8]>, I: IntoIterator<Item=P>

Creates lookup tables to efficiently search for the given needles.

The order of needles is significant, since all Matches returned from this TwoByteWM will include an index into needles saying which needle matched.

fn find_from<P>(&self, haystack: P, offset: usize) -> Option<Match> where P: AsRef<[u8]>

Searches for a single match, starting from the given byte offset.

fn find<'a, 'b>(&'a self, haystack: &'b str) -> Matches<'a, 'b>

Returns an iterator over non-overlapping matches.